Pup 0.8.6: Correct some issues with automatic heading hold
The generic autopilot helper calculates heading-bug-error-deg using
a heading direct from the simulation, rather than a heading that
simulates drift in the directional gyro. The automatic heading hold
is based on readings from the gyro and was confused by the calculated
heading-bug-error-deg. Fix by calculating in the autopilot.

Removed generic autopilot helpers, which are not used.

Modify the trigger for automatic heading hold so that it engages when
the autopilot engages, the heading is stable and the turn select is
neutral. The heading hold was only engaging when the turn select was
returned to neutral.

git-svn-id: http://svn.code.sf.net/p/flightgear/fgaddon/trunk/Aircraft/Beagle-Pup@1626 9eb9bc84-db9d-4e05-893b-26dd41ac94a6